requirements.txt
# Core dependencies
numpy>=1.20.0
opencv-python>=4.5.0
matplotlib>=3.3.0
PyYAML>=5.4.0

# G-code and CNC
pygcode>=0.2.1
pyserial>=3.5
svgpathtools>=1.4.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0
pytest-mock>=3.6.0

# Development tools
black>=22.0.0
mypy>=0.910
flake8>=4.0.0

# Optional UI backends (uncomment as needed)
# PyQt6>=6.2.0
# PySide6>=6.2.0
# fastapi>=0.70.0
# uvicorn>=0.15.0